Remove test due to js-dom
Based on this stack overflow (https://stackoverflow.com/questions/60495903/testing-react-contenteditable-with-react-testing-library+&cd=1&hl=en&ct=clnk&gl=uk) which points to two discussions testing-library/dom-testing-library#235 and jsdom/jsdom#1670

It’s not possible to simulate events on a contenteditable div within a js-dom context. This is a limit of js-dom itself. This is a known issue since 2016.

To save the test, the following was reviewed - modifying the test to click the button but the value of the coverLetter would still be null as it requires that DOM event -> dispatchTranscation -> props.onChange to modify the state of coverLetter.